bump 20.08.18
- Fixed Cache Log -> change to native file_put_contents instead of put() to avoid early unlock result to truncate.
- Fixed Filesystem::put() -> add blocking option to avoid early unlock.
- Fixed dc_save() -> invalid conditional for is_data_updated().
- Fixed skip_stats() -> add checking for ignored groups.
- Fixed unlink -> add checking is_file to avoid php warning and make query-monitor happy.
- Fixed Canopt::setlock() -> set file permission if write true.
- Added DOCKET_CACHE_IGNORED_PRECACHE constant to exclude group:key from precaching.
- Added DOCKET_CACHE_IGNORED_GROUPKEY constant to exclude group:key from persistent cache.
- Added CLI command "unlock" to clear all lock files.

bump 20.08.17

- Fixed CronAgent, woocommerce -> get_cart - not be called before the wp_loaded action.
- Fixed WP_Object_Cache::$cache_hits, WP_Object_Cache::$cache_misses -> hit rate.

+++ includes/src/Event.php
@@ -189,7 +189,7 @@ final class Event
 
                 $fn = $object->getFileName();
                 $fs = $object->getSize();
-                $fm = time() + 120;
+                $fm = time() + 300;
                 $ft = filemtime($fx);
 
                 if ($fm >= $ft && (0 === $fs || 'dump_' === substr($fn, 0, 5))) {
@@ -198,6 +198,8 @@ final class Event
                     ++$collect->clean;
                     continue;
                 }
+
+                $domaxttl = false;
 
                 $data = $this->plugin->cache_get($fx);
                 if (false !== $data) {
@@ -217,6 +219,8 @@ final class Event
                             --$cnt;
                             ++$collect->clean;
                             ++$collect->maxttl_c;
+
+                            $domaxttl = true;
                             continue;
                         }
                     }
@@ -231,7 +235,7 @@ final class Event
                 }
                 unset($data);
 
-                if ($maxttl > 0 && $ft < $maxttl) {
+                if (!$domaxttl && $maxttl > 0 && $ft < $maxttl) {
                     $this->plugin->unlink($fx, true);
                     --$cnt;
                     ++$collect->clean;
